Create README.md
This project is part of a Cyber Security Internship task. The goal was to perform a basic network scan using Nmap to find open ports on devices in the local network. It helped me understand how networks can be scanned, what open ports mean, and how they might expose systems to risks. I used Kali Linux running in VirtualBox and saved the scan results for analysis.